Refactored sync into its own folder
